pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
